Make ServerChannel implementation-independent
