+++ src/main/java/de/btobastian/javacord/entity/channel/impl/ImplServerVoiceChannel.java
@@ -6,6 +6,7 @@ import de.btobastian.javacord.ImplDiscordApi;
 import de.btobastian.javacord.entity.DiscordEntity;
 import de.btobastian.javacord.entity.channel.ChannelCategory;
 import de.btobastian.javacord.entity.channel.InternalChannel;
+import de.btobastian.javacord.entity.channel.InternalServerChannel;
 import de.btobastian.javacord.entity.channel.ServerVoiceChannel;
 import de.btobastian.javacord.entity.channel.ServerVoiceChannelUpdater;
 import de.btobastian.javacord.entity.permission.Permissions;
@@ -25,7 +26,7 @@ import java.util.concurrent.ConcurrentHashMap;
 /**
  * The implementation of {@link ServerVoiceChannel}.
  */
-public class ImplServerVoiceChannel implements ServerVoiceChannel, InternalChannel {
+public class ImplServerVoiceChannel implements ServerVoiceChannel, InternalChannel, InternalServerChannel {
 
     /**
      * The discord api instance.
